+++ sumcheck/src/prover_helper.rs
@@ -1,0 +1,8 @@
+mod power_gate;
+mod product_gate;
+mod simd_gate;
+mod sumcheck_gkr_square;
+mod sumcheck_gkr_vanilla;
+
+pub(crate) use sumcheck_gkr_square::SumcheckGkrSquareHelper;
+pub(crate) use sumcheck_gkr_vanilla::SumcheckGkrVanillaHelper;
